+++ vue-examples/src/components/FilterAcceptOnEnter.vue
@@ -7,7 +7,6 @@
 
       <div class="col-sm-10">
         <MultipleSelect
-          single
           :options="options"
         >
           <option value="1">
@@ -114,6 +113,7 @@
 
       <div class="col-sm-10">
         <MultipleSelect
+          multiple
           :options="options"
         >
           <option value="1">