+++ vue-examples/src/components/ShowClear.vue
@@ -7,7 +7,6 @@
 
       <div class="col-sm-10">
         <MultipleSelect
-          single
           :options="options"
         >
           <option value="1">
@@ -57,7 +56,6 @@
 
       <div class="col-sm-10">
         <MultipleSelect
-          single
           :options="options"
         >
           <optgroup label="Group 1">
@@ -104,6 +102,7 @@
 
       <div class="col-sm-10">
         <MultipleSelect
+          multiple
           :options="options"
         >
           <option value="1">
@@ -153,6 +152,7 @@
 
       <div class="col-sm-10">
         <MultipleSelect
+          multiple
           :options="options"
         >
           <optgroup label="Group 1">