+++ vue-examples/src/components/SingleRadio.vue
@@ -7,7 +7,6 @@
 
       <div class="col-sm-10">
         <MultipleSelect
-          single
           :options="options"
         >
           <option value="1">
@@ -33,7 +32,6 @@
 
       <div class="col-sm-10">
         <MultipleSelect
-          single
           :options="options"
         >
           <optgroup label="Group 1">